ProjectExplorerView: relocate file operations
